+++ app/assets/components/FooterContent.vue
@@ -1,5 +1,3 @@
 <template>
-    <UFFooterContent>
-        <div>{{ $t('COPYRIGHT', { year: new Date().getFullYear() }) }}</div>
-    </UFFooterContent>
+    <div>{{ $t('COPYRIGHT', { year: new Date().getFullYear() }) }}</div>
 </template>