+++ app/assets/layouts/LayoutPage.vue
@@ -1,13 +1,27 @@
 <script setup lang="ts">
-import NavBar from '../components/NavBar.vue'
+import NavBarContent from '../components/NavBarContent.vue'
+import SideBarContent from '../components/SideBarContent.vue'
 import FooterContent from '../components/FooterContent.vue'
+import { useConfigStore } from '@userfrosting/sprinkle-core/stores'
+const config = useConfigStore()
 </script>
 
 <template>
-    <NavBar />
+    <UFNavBar :title="config.get('site.title')" :to="{ name: 'home' }">
+        <NavBarContent />
+    </UFNavBar>
+
+    <!-- Sidebar for small screens -->
+    <UFSideBar class="uk-hidden@m">
+        <SideBarContent />
+    </UFSideBar>
+
     <UFMainContent>
         <UFHeaderPage />
         <RouterView />
     </UFMainContent>
-    <FooterContent />
+
+    <UFFooter>
+        <FooterContent />
+    </UFFooter>
 </template>