+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [6.0.0-beta.7](https://github.com/userfrosting/UserFrosting/compare/6.0.0-beta.6...6.0.0-beta.7)
 - [Docker] Update tp PHP 8.4
+- Add mobile navigation
+- Restructure Navbar/Sidebar/Footer Content files
 
 ## [6.0.0-beta.6](https://github.com/userfrosting/UserFrosting/compare/6.0.0-beta.5...6.0.0-beta.6)
 - Remove Limax from optimizeDeps